[compiler] Prune dependencies that are only used by useRef or useState
Summary: @jmbrown215 recently had an observation that the arguments to useState/useRef are only used when a component renders for the first time, and never afterwards. This is a useful observation for a couple reasons: we can skip more computation that we previously could, with reactive blocks that previously recomputed values when inputs changed now only ever computing them on the first render.

In addition, this is helpful for the debugging tools we're looking into. We'd like to detect when rules of react have been violated, and one way to do so is to detect when a would-be memoized value has changed between renders (see a diff later in this stack for more), but this analysis is noisy if (as seems to be the case in practice) the argument to useState may change between renders. Since such cases are "safe," if ill-advised, this diff will clean up that noise.